+++ include/mujoco/mujoco.h
@@ -465,6 +465,10 @@ MJAPI int mj_stateSize(const mjModel* m, unsigned int sig);
 // Get state.
 MJAPI void mj_getState(const mjModel* m, const mjData* d, mjtNum* state, unsigned int sig);
 
+// Extract a subset of components from a state previously obtained via mj_getState.
+MJAPI void mj_extractState(const mjModel* m, const mjtNum* src, unsigned int srcsig,
+                           mjtNum* dst, unsigned int dstsig);
+
 // Set state.
 MJAPI void mj_setState(const mjModel* m, mjData* d, const mjtNum* state, unsigned int sig);